move init messages
